Improve handling of nulls and start and end of data

+++ lamplighter-model/src/main/resources/com/intrbiz/lamplighter/data/round_time.sql
@@ -2,5 +2,5 @@ CREATE OR REPLACE FUNCTION lamplighter.round_time(p_time TIMESTAMP WITH TIME ZON
 RETURNS TIMESTAMP WITH TIME ZONE 
 LANGUAGE SQL 
 AS $$ 
- SELECT to_timestamp((floor(extract('epoch' FROM $1) / ($2/1000))::BIGINT) * ($2/1000)) 
+ SELECT to_timestamp((floor(extract('epoch' FROM $1) / ($2/1000)::REAL)::BIGINT) * ($2/1000)) 
 $$;
